backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Celery (Task Queue)
celery==5.3.6
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jwt==4.1.0
cryptography==42.0.1

# HTTP & API Clients
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Google Ads API
google-ads==23.1.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# Meta Ads API
facebook-business==19.0.0

# ML & Data Science
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.2.0
numpy==1.26.3
pandas==2.2.0
joblib==1.3.2

# AI & LLM
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.1.1
openai==1.10.0
anthropic==0.18.0
tiktoken==0.6.0

# Vector Store (for AI Copilot RAG)
chromadb==0.4.22
sentence-transformers==2.3.1

# Report Generation
reportlab==4.0.9
openpyxl==3.1.2
jinja2==3.1.3

# Utilities
python-dateutil==2.8.2
pytz==2024.1
pyyaml==6.0.1
python-dotenv==1.0.1

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
sentry-sdk[fastapi]==1.40.1

# Logging
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
faker==22.6.0
